requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
sqlalchemy==2.0.20
pydantic==2.3.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
redis==4.6.0
streamlit==1.26.0
requests==2.31.0
websocket-client==1.6.3
python-multipart==0.0.6
